--- a/typed/Makefile
+++ b/typed/Makefile
@@ -1,0 +1,34 @@
+all:
+
+install:
+	go install ./...
+
+test:
+	go test ./...
+
+gen:
+	go generate
+	
+clean:
+	rm -rf s_* sp_* p_*
+	rm -rf int
+	rm -rf int8
+	rm -rf int16
+	rm -rf int32
+	rm -rf int64
+	rm -rf uint
+	rm -rf uint8
+	rm -rf uint16
+	rm -rf uint32
+	rm -rf uint64
+	rm -rf uintptr
+	rm -rf pointer
+	rm -rf struct
+	rm -rf float32
+	rm -rf float64
+	rm -rf complex64
+	rm -rf complex128
+	rm -rf byte
+	rm -rf rune
+	rm -rf string
+	rm -rf iface