… examples/generic/mailbox_int/mailbox_test.go
… typed/byte/mailbox_test.go
… typed/complex128/mailbox_test.go
… typed/complex64/mailbox_test.go
… typed/float32/mailbox.go
… typed/float32/mailbox_test.go
… typed/float64/mailbox.go
… typed/float64/mailbox_test.go
… typed/iface/mailbox.go
… typed/iface/mailbox_test.go
… typed/int/mailbox.go
… typed/int/mailbox_test.go
… typed/int16/mailbox.go
… typed/int16/mailbox_test.go
… typed/int32/mailbox.go
… typed/int32/mailbox_test.go
… typed/int64/mailbox.go
… typed/int64/mailbox_test.go
… typed/int8/mailbox.go
… typed/int8/mailbox_test.go
… typed/p_byte/mailbox.go
… typed/p_byte/mailbox_test.go
… typed/p_complex128/mailbox.go
… typed/p_complex128/mailbox_test.go
… typed/p_complex64/mailbox.go
… typed/p_complex64/mailbox_test.go
… typed/p_float32/mailbox.go
… typed/p_float32/mailbox_test.go
… typed/p_float64/mailbox.go
… typed/p_float64/mailbox_test.go