@xlab 2016-11-08
I used `-benchtime 10s` because otherwise there is some noise in `BenchmarkMailbox-4`, once time it's 2x faster, another time it's 95% close to built-in channels.