@itsmontoya 2016-11-09
You know, the testing inconsistencies worry me a bit. I think it's an issue related to OSX though. I'm not seeing variances at all on my side (Arch Linux). All in all, this pull request was very well done. Thank you for including the before and after benchmarks.

Can we schedule some time to discuss a couple questions I have about the generics?

Hoping to have this merged by EOW (just depends on both of our schedules for communication)