@itsmontoya 2016-11-09
Some benchstat comparisons
```
name            old time/op    new time/op    delta
Mailbox-4         86.2ns ± 0%    54.2ns ± 2%   ~             (p=0.100 n=3+3)
BatchMailbox-4    10.4µs ± 0%     9.0µs ± 2%   ~             (p=0.100 n=3+3)

name            old alloc/op   new alloc/op   delta
Mailbox-4          8.00B ± 0%    0.00B ±NaN%   ~             (p=0.100 n=3+3)
BatchMailbox-4    0.00B ±NaN%    0.00B ±NaN%   ~     (all samples are equal)

name            old allocs/op  new allocs/op  delta
Mailbox-4           1.00 ± 0%     0.00 ±NaN%   ~             (p=0.100 n=3+3)
BatchMailbox-4     0.00 ±NaN%     0.00 ±NaN%   ~     (all samples are equal)
```

Hell yea @xlab! :+1: